requirements.txt
gunicorn==20.0.4
SQLAlchemy==1.3.16
flask-sqlalchemy==2.4.1
python-dotenv==0.13.0
psycopg2-binary==2.8.5